+++ src/cmdliner_eval.mli
@@ -9,11 +9,17 @@
 
 type 'a eval_ok = [ `Ok of 'a | `Version | `Help ]
 type eval_error = [ `Parse | `Term | `Exn ]
+type 'a eval_exit = [ `Ok of 'a  | `Exit of Cmdliner_info.Exit.code ]
 
 val eval_value :
   ?help:Format.formatter -> ?err:Format.formatter -> ?catch:bool ->
   ?env:(string -> string option) -> ?argv:string array -> 'a Cmdliner_cmd.t ->
   ('a eval_ok, eval_error) result
+
+val eval_value' :
+  ?help:Format.formatter -> ?err:Format.formatter -> ?catch:bool ->
+  ?env:(string -> string option) -> ?argv:string array ->
+  ?term_err:int -> 'a Cmdliner_cmd.t -> 'a eval_exit
 
 val eval_peek_opts :
   ?version_opt:bool -> ?env:(string -> string option) ->