+++ CHANGES.md
@@ -8,6 +8,7 @@
 - In manpage specifications use `$(iname)` in the default 
   introduction of the `ENVIRONMENT` section. Follow up to 
   #168.
+- Add `Cmd.eval_value'` a variation on `Cmd.eval_value`.
 
 v1.2.0 2023-04-10 La Forclaz (VS)
 ---------------------------------